Time: 0 ms (100.00%), Space: 6 MB (24.97%) - LeetHub

+++ 342-power-of-four/342-power-of-four.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPowerOfFour(int n) {
         if(n < 1) return false;
-        long long int set = 0, num0 = 0;
+        int set = 0, num0 = 0;
         while(n){if(n&1) set++; else num0++; n = n>>1;}
         if(set == 1 and num0%2 == 0) return true; else return false;
     }
